+++ packages/dart_firebase_admin/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased patch
+
+- Bump intl to `0.20.0`
+- Added Transaction support (thanks to @evandrobubiak)
+
 ## 0.4.0 - 2024-09-11
 
 - Added `firestore.listCollections()` and `doc.listCollections()`